[Chore] #136 - 파일 분리 형식 변경
