+++ CERTI-iOS/Domain/Interfaces/Repositories/AcquisitionRepository.swift
@@ -1,0 +1,10 @@
+//
+//  AcquisitionRepository.swift
+//  CERTI-iOS
+//
+//  Created by 김나연 on 8/16/25.
+//
+
+protocol AcquisitionRepository {
+    func appendAcquisition(id: Int) async throws
+}
