+++ CERTI-iOS/Domain/Interfaces/Repositories/HomeRepository.swift
@@ -1,0 +1,10 @@
+//
+//  HomeRepository.swift
+//  CERTI-iOS
+//
+//  Created by 김나연 on 8/16/25.
+//
+
+protocol HomeRepository {
+    func appendPreCertification(id: Int) async throws
+}
